+++ SwitchDash/Engine/src/main/java/es/ucm/fdi/switchdash/engine/Graphics.java
@@ -1,7 +1,5 @@
 package es.ucm.fdi.switchdash.engine;
 
-import java.io.IOException;
-
 import es.ucm.fdi.switchdash.engine.utils.MyRect;
 
 /**
@@ -26,14 +24,14 @@ public interface Graphics
      * @param x the destiny x
      * @param y the destiny y
      */
-    public void drawImage(Image image, int x, int y, int alpha);
+    public void drawImage(Image image, int x, int y, float alpha);
 
     /**
      *
      * @param image
      * @param dest
      */
-    void drawImage(Image image, MyRect dest, int alpha);
+    void drawImage(Image image, MyRect dest, float alpha);
 
     /**
      *
@@ -43,16 +41,16 @@ public interface Graphics
      * @param destWidth
      * @param destHeight
      */
-    void drawImage(Image image, int destX, int destY, int destWidth, int destHeight, int alpha);
+    void drawImage(Image image, int destX, int destY, int destWidth, int destHeight, float alpha);
 
     /**
      * Draws a given part of the image.
-     *
-     * @param image the image to draw
+     *  @param image the image to draw
      * @param source the source rect from the image which will be drawn
      * @param dest the dest rect where will be drawn
+     * @param alpha
      */
-    void drawImage(Image image, MyRect source, MyRect dest, int alpha);
+    void drawImage(Image image, MyRect source, MyRect dest, float alpha);
 
 
     /**
